feat: refactor webpack config
