+++ bitwarden_license/bit-browser/tsconfig.json
@@ -1,9 +1,19 @@
 {
   "extends": "../../apps/browser/tsconfig",
-  "files": [],
   "include": [
     "src",
+
     "../../apps/browser/src/**/*.d.ts",
+
+    "../../apps/browser/src/autofill/content/*.ts",
+    "../../apps/browser/src/autofill/fido2/content/*.ts",
+    "../../apps/browser/src/autofill/notification/bar.ts",
+    "../../apps/browser/src/autofill/overlay/inline-menu/**/*.ts",
+    "../../apps/browser/src/platform/ipc/content/*.ts",
+    "../../apps/browser/src/platform/offscreen-document/offscreen-document.ts",
+    "../../apps/browser/src/popup/polyfills.ts",
+    "../../apps/browser/src/vault/content/*.ts",
+
     "../../libs/common/src/autofill/constants",
     "../../libs/common/custom-matchers.d.ts"
   ]
