+++ bitwarden_license/bit-browser/webpack.config.js
@@ -1,17 +1,12 @@
-const { AngularWebpackPlugin } = require("@ngtools/webpack");
+const { buildConfig } = require("../../apps/browser/webpack.base");
 
-const webpackConfig = require("../../apps/browser/webpack.config");
-
-const mainConfig = webpackConfig[0];
-const backgroundConfig = webpackConfig[1];
-
-mainConfig.entry["app/main"] = "../../bitwarden_license/bit-browser/src/popup/main.ts";
-mainConfig.plugins[mainConfig.plugins.length - 1] = new AngularWebpackPlugin({
-  tsconfig: "../../bitwarden_license/bit-browser/tsconfig.json",
-  entryModule: "bitwarden_license/bit-browser/src/popup/app.module#AppModule",
-  sourceMap: true,
+module.exports = buildConfig({
+  popup: {
+    entry: "../../bitwarden_license/bit-browser/src/popup/main.ts",
+    entryModule: "../../bitwarden_license/bit-browser/src/popup/app.module#AppModule",
+  },
+  background: {
+    entry: "../../bitwarden_license/bit-browser/src/platform/background.ts",
+  },
+  tsConfig: "../../bitwarden_license/bit-browser/tsconfig.json",
 });
-
-backgroundConfig.entry = "./src/platform/background.ts";
-
-module.exports = webpackConfig;
